paxmon: refactoring
